requirements.txt
# ── Core analytics ──────────────────────────────────────────────────────────
google-cloud-bigquery==3.25.0
google-cloud-bigquery-storage==2.25.0
pandas==2.2.2
pyarrow==16.1.0
db-dtypes==1.3.0

# ── dbt ─────────────────────────────────────────────────────────────────────
dbt-core==1.8.7
dbt-bigquery==1.8.2

# ── Statistics / experimentation ────────────────────────────────────────────
scipy==1.13.1
statsmodels==0.14.2
numpy==1.26.4

# ── Dashboard ───────────────────────────────────────────────────────────────
streamlit==1.36.0
plotly==5.22.0
altair==5.3.0

# ── Utilities ───────────────────────────────────────────────────────────────
python-dotenv==1.0.1
pyyaml==6.0.1
rich==13.7.1
